xop: add _mm_cmov_si128, infrastructure for XOP ISA extension
XOP was an AMD ISA extension which Intel never adopted.  It is present
on heavy-equipment CPUs (bulldozer, piledriver, etc.), but no longer
supported on Zen.  It never really attained widespread use, but this
could still be useful.

Somewhat surprisingly (at least to me), Intel's SDE supports XOP, so
testing is not particularly difficult.  That said, I would very much
like to verify this on a CPU which actually supports it in hardware;
I'll try to find one after the implementations are all done.